Move ML Job predicates to common folder
These are pure functions and used on both the client and server.
